requirements.txt
pydantic~=2.9.1
pydantic-settings~=2.5.2
fastapi~=0.114.2
uvicorn~=0.30.6
SQLAlchemy~=2.0.34
python-jose~=3.3.0
passlib~=1.7.4
alembic~=1.13.2